+++ src/commands/index.ts
@@ -6,6 +6,11 @@ import epoch from './epoch';
 import killProcess from './kill';
 import update from './update';
 import capabilities from './capabilities';
+import id from './id';
+import hash from './hash';
+import encode from './encode';
+import random from './random';
+import port from './port';
 
 export {
   cts,
@@ -16,4 +21,9 @@ export {
   killProcess,
   update,
   capabilities,
+  id,
+  hash,
+  encode,
+  random,
+  port,
 };